Declare uint(256) as uint256 explicitly

+++ contracts/vestings/VestingBasic.sol
@@ -12,19 +12,19 @@ contract VestingBasic is AccessControl {
 
     address public vestingToken;
     address public vestingCollector;
-    uint public unlockTime; // it must be future, imagine that -> 13rd March 11h am UTC -> 1678705200
+    uint256 public unlockTime; // it must be future, imagine that -> 13rd March 11h am UTC -> 1678705200
 
     Vesting[] public vestingSchedule;
-    uint public nextVestingPeriod;
+    uint256 public nextVestingPeriod;
 
     struct Vesting {
-        uint when; // timestamp of when the vesting item is able to be claimed
-        uint amount; // amount of tokens to withdraw, in Wei
+        uint256 when; // timestamp of when the vesting item is able to be claimed
+        uint256 amount; // amount of tokens to withdraw, in Wei
     }
 
-    event Withdrawal(uint when, uint amount);
+    event Withdrawal(uint256 when, uint256 amount);
 
-    constructor(address _vestingToken, address _vestingCollector, uint _unlockTime) {
+    constructor(address _vestingToken, address _vestingCollector, uint256 _unlockTime) {
         require(_vestingToken != address(0), "Vesting token should be a valid address");
         require(_vestingCollector != address(0), "Vesting collector should be a valid address");
         require(block.timestamp < _unlockTime, "Unlock time should be in the future");
@@ -41,10 +41,10 @@ contract VestingBasic is AccessControl {
     function withdraw() public onlyRole(VESTER_ROLE) {
         require(block.timestamp >= unlockTime, "Vesting schedule should be after unlockTime");
         require(nextVestingPeriod < vestingSchedule.length, "All vesting periods have been claimed");
-        uint claimableAmount;
+        uint256 claimableAmount;
 
         // Foreach vestingSchedule is existing in the array
-        for (uint i = nextVestingPeriod; i < vestingSchedule.length; i++) {
+        for (uint256 i = nextVestingPeriod; i < vestingSchedule.length; i++) {
             // If the vesting schedule is not vested and is available to vest by timestamp
             if (vestingSchedule[i].when <= block.timestamp) {
                 // Increment the claimable amount
@@ -62,12 +62,12 @@ contract VestingBasic is AccessControl {
         IERC20(vestingToken).safeTransfer(vestingCollector, claimableAmount);
     }
 
-    function setVestingSchedule(uint[] memory when, uint[] memory amount) public onlyRole(DEFAULT_ADMIN_ROLE) {
+    function setVestingSchedule(uint256[] memory when, uint256[] memory amount) public onlyRole(DEFAULT_ADMIN_ROLE) {
         require(block.timestamp < unlockTime, "Setting vesting schedule should be before unlockTime");
         require(vestingSchedule.length == 0, "Setting vesting schedule not permitted after first setup");
         require(when.length == amount.length, "When.length length must be the same as Amount.length");
 
-        for (uint i = 0; i < when.length; i++) {
+        for (uint256 i = 0; i < when.length; i++) {
             Vesting memory currentVesting = Vesting(
                 when[i],
                 amount[i]
